requirements.txt
ipykernel
tensorflow==2.14.1
pillow==10.2.0
tqdm
faiss-cpu
streamlit==1.33.0
from-root
-e .